django/website/pyproject.toml
[tool.poetry]
name = "wagtail-gridsome"
version = "0.1.0"
description = "Wagtail in Gridsome"
authors = ["Oleg Lavrovsky <oleg@spam-me-later.cc>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6"
django-adminactions = "1.*"
django-extensions = "3.*"
django-cors-headers = "3.10.*"
wagtail-grapple = "^0.12"
channels = "^3"
wagtail = "^2"
dj_database_url = "^0.5.0"
gunicorn = "^20.1.0"
concurrent_log_handler = "^0.9.19"

[tool.poetry.dev-dependencies]
bpython = "^0.21"
django-debug-toolbar = "^3.2"
django-debug-toolbar-template-profiler = "^2.0.2"
Werkzeug = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
